pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openclaw-builder-agent"
version = "0.1.0"
description = "Builder: OpenClaw-native agent implementation for generating and upgrading agents"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "pyyaml>=6.0.1",
  "requests>=2.32.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "pytest-mock>=3.14.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.setuptools.packages.find]
where = ["."]
include = ["builder_agent*"]
